Issue #816
